[Minor] Improve English, use better names and data, simplify SQL (#1265)
### What changes were proposed in this pull request?

In general, it is better to use more realistic data and don't use
generic names ending in numbers. The SQL also needs to match the amount
of data we have in the playground.

### Why are the changes needed?

To provide a slightly more realistic example and make SQL easier to
understand.

Fix: # N/A

### Does this PR introduce _any_ user-facing change?

N/A

### How was this patch tested?

Ran commands locally to test.