[#1419] fix(UI): Fix adding a property of the same name updates the value  (#1622)
**Changes proposed:**

When creating a metalake or editing a metalake, it was allowing users to
enter keys with the same name without notifying the user. So, to solve
this, a little error message underneath the property text field says
that the key is already in use. Also, if they have any duplicate fields
present, it prevents them from creating the meta lake or adding new
fields until the duplicate is resolved.

**Why these changes are needed?**

This change was needed because it reduces the confusion for the user who
may not know the nature of relational databases and that they cannot
have the same keys. So instead of just having the key disappear and
override the previous value, it will properly notify the user that they
need to rename.

Fix #1419 

**Does this PR introduce any user-facing change?**

Yes and no. There is not a significant user change other than it just
helps the user know what is going on and directs them in the correct
way.

**How was this patch tested?**
Locally