[#2457] improve(core): Refactor and reorganize the catalog related codes to be more pluggable (#2482)
### What changes were proposed in this pull request?

This PR refactors the catalog related code a bit in core to make them
more pluggable, and help developers develop a catalog on their own:

1. Create a new `connector` package, and move all the required classes
to this package.
2. Add annotation to this package to show the status of apis.

### Why are the changes needed?

This change will have the users to understand and develop a catalog
easier out of Gravitino.

There still left something that needs to refactor to make it fully
pluggable, like entities, serde and storage. The current refactorings
can only support external catalogs that doesn't require store the
complete metadata in Gravitino.

Fix: #2457 

### Does this PR introduce _any_ user-facing change?

Yes.

### How was this patch tested?

Existing UTs.