[#2478] Improvement(client-java): Separate Java clients for admin and normal user (#2483)
### What changes were proposed in this pull request?

Separate the current Java client into two, one is "GravitinoAdminClient"
which similar to as-is version; the other is "GravitinoClient" which has
a "metalake_name" as input parametor, and then user mainly works on
catalogs.

### Why are the changes needed?

Today in the Java client, the "GravitinoClient" has all the operations
for metalakes, including list, create, alter and drop, etc. That is
mostly for the system administrator's operation, not friendly for a
normal user. Normal user mainly works on catalogs.

Fix: #2478

### Does this PR introduce _any_ user-facing change?

Yes, the Java client "GravitinoClient.java" has user interface change;
If the user want to operate Metalake, he can switch to use the
"GravitinoAdminClient.java".

### How was this patch tested?

The UT for the GravitinoClient can ensure the function is not broken.