[#2759] feat(server,core): Add service admin and metalake admin (#2758)
### What changes were proposed in this pull request?
Role in the Gravitino has 3 kinds: service admins, metalake admins,
normal roles.
Service admins are responsible for creating metalake admins. They can't
be changed by API. It is a config option.
Metalake admin is used for creating metalake. It can't be bind to any
metalake. So it's hard to unify it with normal users.
This pull request add the management of service admins and metalake
admins.

### Why are the changes needed?

Fix: #2759 

### Does this PR introduce _any_ user-facing change?
Yes, I will add openapi and document.

### How was this patch tested?
New ut.

---------

Co-authored-by: Heng Qin <qqtt@123.com>