[#3895] feat(core): Add a basic tag manager framework for tag system (#3908)
### What changes were proposed in this pull request?

This PR proposes to add a basic `TagManager` framework. The current
framework is not ready to work since it misses the core logic.

### Why are the changes needed?

This subtask adds a basic tag framework without actual logic. The reason
of adding this is to control the PR size to avoid a big PR, since
there're many changes in RDBMS support. If we want to make it complete,
the PR will be very big.

Fix: #3895 

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

The current PR is not ready to work, so tests will be added later on.