[3286] Improvement(ui): row tooltip displays comment (#3882)
### What changes were proposed in this pull request?

In this PR, I've changed the tool tip to show the metalake comment and
catalog comment when hovering over the metalake and catalog in each row
respectively.

### Why are the changes needed?

Before the PR when hovering over the metalake row, it would display the
metalake name which is redundant. Additionally, when hovering over the
catalog name it did not show any tooltip at all.

Fix: #3286 

### Does this PR introduce _any_ user-facing change?

N/A

### How was this patch tested?
I tested this using this the web UI to see if it displays correctly when
the mouse is hovered over.