[#3700] refactor(API): Refactor client side MessagingCatalog to use relative path in NameIdentifier  (#3814)
### What changes were proposed in this pull request?

Currently, in the MessagingCatalog.java, the methods like "loadTopic",
"createTopic", "updateTopic" all need a NameIdentifier parameter, which
needs to be a fully-qualified (metalake.catalog.schema.topic) name. But
the "metalake" and "catalog" are not needed, as they already be provided
when load the catalog. To make the API clear and easier to use, we will
change it to use a relative NameIdentifier object (which is
"schema.topic") as the fileset's ID, so that the user doesn't need to
provide the metalake and catalog names repeatedly.

Please note, this only affects the client side.

### Why are the changes needed?

To make the API simple and easy to understand.

Fix: #3700

### Does this PR introduce _any_ user-facing change?

No behavior change, just method parameter.

### How was this patch tested?

No introduce new class or method, so the change will be covered by all
existing test cases.
All related unit tests and integration tests are updated.