[#4236] feat(core): Supports the post-hook for the managers or dispatcher (#4239)
### What changes were proposed in this pull request?
Supports the post-hook for the managers or dispatcher.
For example, after we create a securable object, we should set an owner
for it. We can add a specific post hook to support it.
This pull request uses Java dynamic proxy mechanism to support this
feature. We only support post-hook now, we can support pre-hook in the
future.

### Why are the changes needed?

Fix: #4236 

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Add a new ut.