[#4973] improvement(client-java/server): Clear `CallerContext` in Java Client and Server when call `getFileLocation` finally (#4974)
### What changes were proposed in this pull request?

Currently, after the client completes the `getFileLocation` request and
the server completes the `getFileLocation` response, the Thread Local
`CallerContext` is not cleaned up. This may cause the next request to
incorrectly report the same information repeatedly when the thread is
reused. We should eventually clean up the `CallerContext`.

### Why are the changes needed?

Fix: #4973 

### How was this patch tested?

Add UTs.