[#5825] improvement(CLI): Update the error message for creating a metalake when the required parameter is missing. (#5851)
### What changes were proposed in this pull request?

Currently, when a metalake is created without specifying the metalake
argument, the message "Cannot parse a null or empty identifier" is
displayed. This message is unclear and lacks helpful guidance for the
user. Update the message to: "! metalake is not defined" to make it more
user-friendly and informative.

### Why are the changes needed?

Fix: #5825 

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

<img width="282" alt="image"
src="https://github.com/user-attachments/assets/cb59089f-e60f-4c12-a781-0a79e449c472"
/>