[#5624] feat(bundles): support ADLS credential provider (#5737)
### What changes were proposed in this pull request?

Add ADLS credential provider

### Why are the changes needed?

Fix: #5624

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Added a unit test and verified successful access to the ADLS container.

![image](https://github.com/user-attachments/assets/f8f14696-73b9-4c07-b047-5b257fc7b4a5)

### Supplementary Information

Chose to use the following versions of Azure libraries:

- `azure-identity` = "1.13.1"
- `azure-storage-file-datalake` = "12.20.0"
- `azure-core-http-okhttp` = "1.12.0"

Instead of the latest versions because, although the official
documentation states support for Java 8 and later, the latest versions
appear to have been compiled with Java 21. This caused compilation
issues in the Gravitino environment. Therefore, downgraded and tested to
find the latest usable versions.