[#6694] improvement: Avoid binding type `PATH` to `FILESET` (#6698)
### What changes were proposed in this pull request?

Avoid binding type `PATH` to `FILESET`. When we store a Hive table,
first, we will store the metadata in the MySQL, and then we will store
the data in the HDFS location. For Hive, there is a default location for
the cluster level. We can also specify other locations for schema and
table levels. So a path won't only be fileset.

### Why are the changes needed?

Fix: #6698

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

CI passed.