[#5202] feat(client-python): Support Gravitino Type Serdes - serializer (#6903)
### What changes were proposed in this pull request?

This is the second part (totally 4 planned) of implementation to the
following classes from Java to support Column and its default value,
including:

- JsonUtils
- TypeSerializer

The `TypeSerializer` will be used in the incoming `ColumnDTO`
implementation to serialize `data_type` field.

### Why are the changes needed?

We need to support Column and its default value in python client.

#5202 

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit tests

---------

Signed-off-by: George T. C. Lai <tsungchih.hd@gmail.com>