[#5202] feat(client-python): Support Gravitino Type Serdes - deserializer (#7241)
### What changes were proposed in this pull request?

This is the second part (totally 4 planned) of implementation to the
following classes from Java to support Column and its default value,
including:

- JsonUtils: `readDataType()` and its related methods to deserialize
various Gravitino Types.
- TypeDeserializer

The `TypeDeserializer` will be used in the incoming `ColumnDTO`
implementation to deserialize `data_type` field.

### Why are the changes needed?

We need to support Column and its default value in python client.

#5202 

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit tests.

---------

Signed-off-by: George T. C. Lai <tsungchih.hd@gmail.com>