[#7251] Make `catalogCache` in `CatalogManager` as a instance field not a static value. (#7297)
### What changes were proposed in this pull request?

Changed the catalogCache to be an instance variable instead of static.
This was done to avoid creating the static var from within the
constructor.
This cleaned up some unit tests and places where constructor was first
called to initialize the cache, and then the cache was accessed in a
static manner.

### Why are the changes needed?

Code improvements.

Fix: #7251 

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

./gradlew clean build