[#7975] improvement(validateFieldExistence): empty or null check for fieldName (#7990)
### What changes were proposed in this pull request?

This PR adds an empty or null check for the fieldName array for
validateFieldExistence function.

(Please outline the changes and how this PR fixes the issue.)

### Why are the changes needed?

(Please clarify why the changes are needed. For instance,
  1. If you propose a new API, clarify the use case for a new API.
  2. If you fix a bug, describe the bug.)

Fix: #7975 

### Does this PR introduce _any_ user-facing change?
No

(Please list the user-facing changes introduced by your change,
including
  1. Change in user-facing APIs.
  2. Addition or removal of property keys.)

### How was this patch tested?

(Please test your changes, and provide instructions on how to test it:
1. If you add a feature or fix a bug, add a test to cover your changes.
2. If you fix a flaky test, repeat it for many times to prove it works.)

---------

Co-authored-by: Sambhavi Pandey <sambhavi.pandey@aexp.com>