[#8011] Improvement(LockManager): Corrected code in LockManager.java (#8043)
### What changes were proposed in this pull request?

This PR fixes a bug in the checkDeadLock() function inside
core/src/main/java/org/apache/gravitino/lock/LockManager.java.
The code is modified to correctly format each thread's timestamp inside
the forEach loop.

### Why are the changes needed?

Previously, the deadlock logging in LockManager was trying to format the
entire map instead of the actual timestamp, which was leading to an
IllegalArgumentException when checking for long-held locks.

Fix: #8011

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Locally, with a unit test provided in the original issue. Since the fix
basically amounts to fixing a typo rather than changing any core
functionality, I've opted to not include the test with the PR.